LayoutTests/media/modern-media-controls/localized-strings/replaced-string.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/en.lproj/modern-media-controls-localized-strings.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/main.js" type="text/javascript"></script>
<script type="text/javascript">

description("Testing that we can replace strings using UIString.");

UIStrings["Test replaced string"] = "Value %s should be printed here";

shouldBeEqualToString("UIString('Test replaced string', 15)", "Value 15 should be printed here");

</script>
<script src="../../../resources/js-test-post.js"></script>
